Split out item categories.
