+++ src/app/loadout/loadout-ui/ItemCategorySection.m.scss
@@ -1,0 +1,128 @@
+@import '../../variables.scss';
+
+.itemCategory {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  @include vertical-space-children(var(--item-margin));
+
+  :global(.stat-row) {
+    justify-content: space-between;
+  }
+
+  :global(.stat-bars) {
+    margin-top: 0;
+  }
+}
+
+.itemsInCategory {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, var(--item-size));
+  gap: var(--item-margin);
+  width: 100%;
+}
+
+.itemBucket {
+  width: var(--item-size);
+  @include vertical-space-children(var(--item-margin));
+}
+
+.items {
+  display: grid;
+  gap: 2px;
+  grid-template-columns: 1fr 1fr;
+  min-height: unset !important;
+}
+
+.empty {
+  height: calc(var(--item-size) + #{$badge-height} - #{3 * $item-border-width});
+  border: 1px solid white;
+}
+
+.categoryWeapons {
+  width: calc(3 * var(--item-size) + 2 * var(--item-margin));
+}
+.categoryArmor {
+  width: calc(5 * var(--item-size) + 4 * var(--item-margin));
+}
+.categoryGeneral {
+  width: calc(2 * var(--item-size) + 1 * var(--item-margin));
+}
+
+.equipped {
+  composes: sub-bucket from global;
+  width: 100%;
+  padding: 0;
+}
+
+.unequipped {
+  composes: sub-bucket from global;
+  width: 100%;
+  padding: 0;
+  --item-icon-size: calc(0.5 * var(--item-size) - 2px);
+  :global(.item) {
+    --item-size: var(--item-icon-size);
+  }
+}
+
+.loParams {
+  max-width: 100%;
+  @include vertical-space-children(4px);
+}
+
+.loStats {
+  composes: flexRow from '../../dim-ui/common.m.scss';
+  align-items: center;
+  width: 100%;
+  @include horizontal-space-children(8px);
+}
+
+.loStat {
+  composes: flexRow from '../../dim-ui/common.m.scss';
+  align-items: center;
+  font-weight: bold;
+  font-size: 11px;
+
+  img {
+    height: 14px;
+    width: 14px;
+    margin-right: 2px;
+  }
+}
+
+.loExotic {
+  --item-size: 17px;
+  span {
+    margin-left: 4px;
+  }
+}
+
+.loSpendTier {
+  > div {
+    padding: 0;
+    margin: 0;
+  }
+}
+
+.loQuery {
+  overflow: hidden;
+  max-width: 100%;
+  max-height: 4em;
+  text-overflow: ellipsis;
+  display: -webkit-box;
+  -webkit-line-clamp: 3;
+  -webkit-box-orient: vertical;
+
+  :global(.app-icon) {
+    margin-left: 3px;
+    margin-right: 6px;
+  }
+}
+
+.missingItem {
+  opacity: 0.3;
+}
+
+.placeholder {
+  composes: placeholder from '../Loadouts.m.scss';
+}
