+++ src/app/loadout/loadout-ui/ItemCategorySection.m.scss.d.ts
@@ -1,0 +1,24 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'categoryArmor': string;
+  'categoryGeneral': string;
+  'categoryWeapons': string;
+  'empty': string;
+  'equipped': string;
+  'itemBucket': string;
+  'itemCategory': string;
+  'items': string;
+  'itemsInCategory': string;
+  'loExotic': string;
+  'loParams': string;
+  'loQuery': string;
+  'loSpendTier': string;
+  'loStat': string;
+  'loStats': string;
+  'missingItem': string;
+  'placeholder': string;
+  'unequipped': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
